requirements.txt
flask==3.0.3
flask-cors==4.0.0
pymongo[srv]==4.8.0
python-dotenv==1.0.1


